use mulitiple thread to check port

+++ vpngate_list_auto.py
@@ -4,14 +4,11 @@ import re
 import base64
 import socket
 import os, glob, sys, shutil
+import threading
+import Queue
+import logging
 
 vpn_list = 'http://enigmatic-scrubland-4484.herokuapp.com/'
-
-# get serer list from list url
-ua = urllib2.Request(vpn_list)
-ua.add_header('User-agent', 'Mozilla/5.0')
-
-res = urllib2.urlopen(ua)
 
 def tcp_port_is_open(ip, port) :
     s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -22,6 +19,9 @@ def tcp_port_is_open(ip, port) :
         return True
     else :
         return False
+
+def udp_port_is_open(ip, port):
+    pass
 
 
 def save_config_file(result) :
@@ -36,13 +36,15 @@ def save_config_file(result) :
         else:
             config_path = 'C:\Program Files\OpenVPN\config'
     else:
-        loging.error("unkonw os type")
+        logging.error("unkonw os type")
 
 
     # rm old config files
     os.chdir(config_path) # will auto die if chdir fail
     backup_dir = 'vpngate_old'
-    shutil.rmtree(backup_dir)
+
+    if os.path.exists(backup_dir):
+        shutil.rmtree(backup_dir)
     os.mkdir(backup_dir)
 
     # backup config files
@@ -60,70 +62,123 @@ def save_config_file(result) :
             f.write(server['config'])
             f.close
 
+def get_config_content(vpn_lists):
+    for vpn_list in vpn_lists:
+        ua = urllib2.Request(vpn_list)
+        ua.add_header('User-agent', 'Mozilla/5.0')
+        res = urllib2.urlopen(ua)
 
-if res.getcode() == 200 :
-
-    result = {}
-    print dir(res)
-    #  deal every line , skip comment lines
+        res_code = res.getcode()
+        if res_code == 200:
+            print 'web fetch ok'
+            return res.read()
+        else:
+            logging.warn("vpn_list %s fetch error: %s" % (vpn_list, res_code))
+            continue
     
-    p = re.compile('^\w+')
+    return None
 
-    while True :
-        svr_line = res.readline()
+def parse_config_content(content):
+    '''
+    res is web conntent handle
+    parse all server, put into job queue
+    '''
+    svr_queue = Queue.Queue()
+    p = re.compile('^\w+')
+    p_newline = re.compile('\n')
+
+    for svr_line in re.compile('\n').split(content):
         if svr_line:
             if p.match(svr_line):
                 c = re.compile(',').split(svr_line)
                 ip = c[1]
                 country = c[6]
-                config_base64 = c[-1]
-                config = base64.b64decode(config_base64)
-#                print ip, country, config
-                print ip, country
+                config = base64.b64decode(c[-1])
 
-                # get tcp port from config_file
-                p_tcp = re.compile('^proto tcp', re.MULTILINE)
+                # get protocol, port from config
+                p_proto = re.compile('^proto ((?:tcp)|(?:udp))', re.MULTILINE)
                 p_port = re.compile('^remote [.|\d]+ (\d+)', re.MULTILINE)
-                if p_tcp.search(config) :
+
+                m_proto = p_proto.search(config)
+                if m_proto:
+                    proto = m_proto.group(1)
+
                     m_port = p_port.search(config)
-                    if m_port :
-                        port = int(m_port.group(1)) # 80 is num, '80' is str, it's different betwen perl and python
-#                        print ip, port
-                        if tcp_port_is_open(ip, port) :
-                            print "GOOD: ", ip, port
-                            if not country in result :
-                                result[country] = []
-                            else :
-                                result[country].append({'ip':ip, 'config':config})
-                            
-                        else :
-                            print "TIMEOUT: ", ip, port
+                    if m_port:
+                        port = m_port.group(1)
 
-        else :
+
+                        ## and server info into job queue
+                        svr_queue.put((ip, proto, int(port), config, country))
+          
+                else:
+                    logging.warn('Can"t find proto at like %s' % 'LINENUM')
+
+    return svr_queue          
+                
+
+class WorkerThread(threading.Thread):
+    def __init__(self, svr_queue, result_queue):
+        threading.Thread.__init__(self)
+        self.svr_queue = svr_queue
+        self.result_queue = result_queue
+
+    def run(self):
+        while True:
+            ip, proto, port, config, country = self.svr_queue.get()
+            
+            ava = False
+            if proto == 'tcp':
+                ava = tcp_port_is_open(ip, port)
+            elif proto == 'udp':
+                ava = udp_port_is_open(ip, port)
+            else:
+                logging.warn('ip:%s port:%s proto:%s error' % (ip, port, proto))
+
+            if ava:
+                self.result_queue.put((ip, proto, port, config, country))
+            else:
+                print "timeout ip:%s port %s" % (ip, port)
+            self.svr_queue.task_done()
+            print "done ip:%s, port:%s" % (ip, port)
+
+def check_ava_svr(svr_queue):
+    '''
+    muiltiple thread 
+    '''
+    result_queue = Queue.Queue()
+
+    t_num = 20
+    for i in range(t_num):
+        t = WorkerThread(svr_queue, result_queue)
+        t.setDaemon(True)
+        t.start()
+
+    svr_queue.join()
+
+    return result_queue
+
+
+if __name__ == '__main__':
+    vpn_lists = (vpn_list,)
+    web_res = get_config_content(vpn_lists)
+
+    svr_queue = parse_config_content(web_res)
+    result_queue = check_ava_svr(svr_queue)
+
+    result = {}
+    while True:
+        try:
+            ip, proto, port, config, country = result_queue.get(block=False)
+            print "Good: ip:%s port:%s proto:%s" % (ip, port, proto)
+        except Exception as e:
+            logging.info(e)
             break
-    
-    save_config_file(result) 
-    '''
-   #print result
 
-    # rm old file
-    config_path = 'C:\Program Files\OpenVPN\config'
-    os.chdir(config_path)
-    for conf in glob.glob('vpngate*.ovpn') :
-        os.remove(conf)
+        if not country in result:
+            result[country] = []
+        result[country].append({'ip':ip, 'port':port, 'proto':proto, 'config':config})
 
-    # write to file, every country limit 3 server
-    for country in result :
-        for server in result[country][0:3] :
-#            print server
-            file_name = '_'.join(['vpngate', country, server['ip']]) + '.ovpn'
-            print file_name
-            f = open(file_name, 'w')
-            f.write(server['config'])
-            f.close()
+    save_config_file(result)
 
-    '''
-
-else :
-    print res.getcode()
